+++ crates/primitives/receipt/src/from_blockifier.rs
@@ -3,13 +3,16 @@ use blockifier::transaction::{
     account_transaction::AccountTransaction,
     objects::{FeeType, GasVector, HasRelatedFeeType, TransactionExecutionInfo},
     transaction_execution::Transaction,
+    transactions::L1HandlerTransaction,
 };
 use cairo_vm::types::builtin_name::BuiltinName;
+use starknet_core::types::MsgToL2;
 use starknet_types_core::felt::Felt;
+use thiserror::Error;
 
 use crate::{
     DeclareTransactionReceipt, DeployAccountTransactionReceipt, Event, ExecutionResources, ExecutionResult, FeePayment,
-    InvokeTransactionReceipt, L1Gas, MsgToL1, PriceUnit, TransactionReceipt,
+    InvokeTransactionReceipt, L1Gas, L1HandlerTransactionReceipt, MsgToL1, PriceUnit, TransactionReceipt,
 };
 
 fn blockifier_tx_fee_type(tx: &Transaction) -> FeeType {
@@ -29,6 +32,33 @@ fn blockifier_tx_hash(tx: &Transaction) -> Felt {
     }
 }
 
+#[derive(Debug, Error)]
+pub enum L1HandlerMessageError {
+    #[error("Empty calldata")]
+    EmptyCalldata,
+    #[error("From address out of range")]
+    FromAddressOutOfRange,
+    #[error("Invalid nonce")]
+    InvalidNonce,
+}
+
+fn get_l1_handler_message_hash(tx: &L1HandlerTransaction) -> Result<Felt, L1HandlerMessageError> {
+    let (from_address, payload) = tx.tx.calldata.0.split_first().ok_or(L1HandlerMessageError::EmptyCalldata)?;
+
+    let from_address = (*from_address).try_into().map_err(|_| L1HandlerMessageError::FromAddressOutOfRange)?;
+
+    let nonce = tx.tx.nonce.0.to_bigint().try_into().map_err(|_| L1HandlerMessageError::InvalidNonce)?;
+
+    let message = MsgToL2 {
+        from_address,
+        to_address: tx.tx.contract_address.into(),
+        selector: tx.tx.entry_point_selector.0,
+        payload: payload.into(),
+        nonce,
+    };
+    Ok(Felt::from_bytes_le(message.hash().as_bytes()))
+}
+
 pub fn from_blockifier_execution_info(res: &TransactionExecutionInfo, tx: &Transaction) -> TransactionReceipt {
     let price_unit = match blockifier_tx_fee_type(tx) {
         FeeType::Eth => PriceUnit::Wei,
@@ -40,6 +70,17 @@ pub fn from_blockifier_execution_info(res: &TransactionExecutionInfo, tx: &Trans
 
     let mut events: Vec<Event> = Vec::new();
     get_events_from_call_info(res.execute_call_info.as_ref(), 0, &mut events);
+
+    let message_hash = match tx {
+        Transaction::L1HandlerTransaction(tx) => match get_l1_handler_message_hash(tx) {
+            Ok(hash) => Some(hash),
+            Err(err) => {
+                log::error!("Error getting l1 handler message hash: {:?}", err);
+                None
+            }
+        },
+        _ => None,
+    };
 
     let messages_sent = res
         .non_optional_call_infos()
@@ -113,16 +154,15 @@ pub fn from_blockifier_execution_info(res: &TransactionExecutionInfo, tx: &Trans
                 execution_result,
             })
         }
-        Transaction::L1HandlerTransaction(_tx) => unimplemented!("l1 handler tx"),
-        // Transaction::L1HandlerTransactionv(tx) => TransactionReceipt::L1Handler(L1HandlerTransactionReceipt {
-        //     transaction_hash,
-        //     actual_fee,
-        //     messages_sent,
-        //     events,
-        //     execution_resources,
-        //     execution_result,
-        //     message_hash: todo!(),
-        // }),
+        Transaction::L1HandlerTransaction(_tx) => TransactionReceipt::L1Handler(L1HandlerTransactionReceipt {
+            transaction_hash,
+            actual_fee,
+            messages_sent,
+            events,
+            execution_resources,
+            execution_result,
+            message_hash: message_hash.unwrap(), // it's a safe unwrap because it would've panicked earlier if it was Err
+        }),
     }
 }